+++ app/Http/Controllers/PolicyController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Policy;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\File;
 
 class PolicyController extends Controller
@@ -28,7 +29,17 @@ class PolicyController extends Controller
      */
     public function create()
     {
-        return view('admin.policies.create');
+        $authuser = Auth::user();
+        if ($authuser->hradmin == "yes")
+        {
+            return view('admin.policies.create');
+        }
+        else
+        {
+            abort(403);
+        }
+
+        
     }
 
     /**
@@ -39,27 +50,36 @@ class PolicyController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
+        $authuser = Auth::user();
+        if ($authuser->hradmin !== "yes")
+        {
+            abort(403);
+        }
+        else
+        {
+            $request->validate([
 
-            'name' => 'required|unique:policies,name',
-            'desc',
-            'created_date' => 'required',
-            'lastupdate_date' => 'required',
-            'file' => 'required',
-        ]);
-
-        $path = $request->file('file')->storeAs('public/files', $request->name . '.pdf');
-
-        $policy = new Policy();
-        $policy->name = $request->name;
-        $policy->desc = $request->desc;
-        $policy->created_date = $request->created_date;
-        $policy->lastupdate_date = $request->lastupdate_date;
-        $policy->path = $path;
-        $policy->save();
-
-        $policy = Policy::all();
-        return view('admin.policies.index', ['policies' => $policy]);
+                'name' => 'required|unique:policies,name',
+                'desc',
+                'created_date' => 'required',
+                'lastupdate_date' => 'required',
+                'file' => 'required',
+            ]);
+    
+            $path = $request->file('file')->storeAs('public/files', $request->name . '.pdf');
+    
+            $policy = new Policy();
+            $policy->name = $request->name;
+            $policy->desc = $request->desc;
+            $policy->created_date = $request->created_date;
+            $policy->lastupdate_date = $request->lastupdate_date;
+            $policy->path = $path;
+            $policy->save();
+    
+            $policy = Policy::all();
+            return view('admin.policies.index', ['policies' => $policy]);
+        }
+        
     }
 
     /**
@@ -81,8 +101,16 @@ class PolicyController extends Controller
      */
     public function edit(Policy $policy)
     {
-
-        return view('admin.policies.edit', ['policy' => $policy]);
+        $authuser = Auth::user();
+        if ($authuser->hradmin == "yes")
+        {
+            return view('admin.policies.edit', ['policy' => $policy]);
+        }
+        else
+        {
+            abort(403);
+        }
+        
     }
 
     /**
@@ -119,11 +147,21 @@ class PolicyController extends Controller
      */
     public function destroy(Policy $policy)
     {
-        // File::delete($policy->name);
-        // File::delete(public_path("files/{{$policy->name}} . '.pdf'"));
-        $file_path = public_path() . '/storage/files/' . $policy->name . '.pdf';
-        unlink($file_path);
-        $policy->delete();
-        return redirect()->route('admin.policies.index');
+
+        $authuser = Auth::user();
+        if ($authuser->hradmin == "yes")
+        {
+             // File::delete($policy->name);
+             // File::delete(public_path("files/{{$policy->name}} . '.pdf'"));
+          $file_path = public_path() . '/storage/files/' . $policy->name . '.pdf';
+          unlink($file_path);
+          $policy->delete();
+          return redirect()->route('admin.policies.index');
+        }
+        else
+        {
+            abort(403);
+        }
+       
     }
 }
